[BEAM-2236] Move test utilities out of python core
