+++ sdks/python/apache_beam/examples/complete/autocomplete_test.py
@@ -21,7 +21,7 @@ import unittest
 
 import apache_beam as beam
 from apache_beam.examples.complete import autocomplete
-from apache_beam.test_pipeline import TestPipeline
+from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.transforms.util import assert_that
 from apache_beam.transforms.util import equal_to
