+++ sdks/python/apache_beam/examples/complete/tfidf_test.py
@@ -25,7 +25,7 @@ import unittest
 
 import apache_beam as beam
 from apache_beam.examples.complete import tfidf
-from apache_beam.test_pipeline import TestPipeline
+from apache_beam.testing.test_pipeline import TestPipeline
 
 
 EXPECTED_RESULTS = set([
