+++ sdks/python/apache_beam/examples/cookbook/combiners_test.py
@@ -27,7 +27,7 @@ import logging
 import unittest
 
 import apache_beam as beam
-from apache_beam.test_pipeline import TestPipeline
+from apache_beam.testing.test_pipeline import TestPipeline
 
 
 class CombinersTest(unittest.TestCase):
