+++ sdks/python/apache_beam/io/textio_test.py
@@ -43,7 +43,7 @@ from apache_beam.io.filebasedsource_test import write_data
 from apache_beam.io.filebasedsource_test import write_pattern
 from apache_beam.io.filesystem import CompressionTypes
 
-from apache_beam.test_pipeline import TestPipeline
+from apache_beam.testing.test_pipeline import TestPipeline
 
 from apache_beam.transforms.util import assert_that
 from apache_beam.transforms.util import equal_to
