+++ sdks/python/apache_beam/testing/data/standard_coders.yaml
