+++ sdks/python/apache_beam/testing/test_pipeline_test.py
@@ -24,7 +24,7 @@ from hamcrest.core.base_matcher import BaseMatcher
 from hamcrest.core.assert_that import assert_that as hc_assert_that
 
 from apache_beam.internal import pickler
-from apache_beam.test_pipeline import TestPipeline
+from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.options.pipeline_options import PipelineOptions
