+++ sdks/python/apache_beam/transforms/create_test.py
@@ -22,7 +22,7 @@ from apache_beam.io import source_test_utils
 
 from apache_beam import Create, assert_that, equal_to
 from apache_beam.coders import FastPrimitivesCoder
-from apache_beam.test_pipeline import TestPipeline
+from apache_beam.testing.test_pipeline import TestPipeline
 
 
 class CreateTest(unittest.TestCase):
