+++ sdks/python/apache_beam/transforms/util_test.py
@@ -20,7 +20,7 @@
 import unittest
 
 from apache_beam import Create
-from apache_beam.test_pipeline import TestPipeline
+from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.transforms.util import assert_that, equal_to, is_empty
