+++ sdks/python/apache_beam/transforms/write_ptransform_test.py
@@ -22,7 +22,7 @@ import unittest
 import apache_beam as beam
 
 from apache_beam.io import iobase
-from apache_beam.test_pipeline import TestPipeline
+from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.transforms.ptransform import PTransform
 from apache_beam.transforms.util import assert_that, is_empty
